configs: add Freescale SABRE Auto board support
SABRE Board for Automotive Infotainment (SABRE Auto, a.k.a. SABRE-AI) is
Freescale's evaluation board based on the i.MX 6 ARM Cortex-A9 applications
processor.

This defconfig is an adaptation of freescale_imx6sabresd_defconfig for SABRE
Auto, and is based on Freescale "official" git repo on git.freescale.com and SW
release 3.0.35_4.1.0.

Signed-off-by: Vincent Stehlé <vincent.stehle@freescale.com>
Cc: Gilles Talis <gilles.talis@gmail.com>
Cc: Peter Korsgaard <jacmet@sunsite.dk>