openocd
This patch adds the management of (almost) all the config options of
openocd 0.8.0. A BR config variable is added for (almost) every adapter
known to openocd and all the dependencies are automatically calculated
from the chosen adapters, so only the necessary libraries are built.
Note that CMSIS_DAP adapter requires hidapi (not libhid) and hidapi
is not actually included in buildroot, so it has been removed.
Also zy1000 adapters are actually broken in openocd (as far as I know)
and have been removed.
The host version of the package enables all the possible adapters and
the related libraries.

Signed-off-by: Claudio Laurita <claudio.laurita@integrazionetotale.it>