configs/lego_ev3: bump kernel to 4.4 series
Bump kernel revision used by lego_ev3_defconfig to the latest tag from
ev3dev, as suggested by Peter. This is a 4.4 series kernel, which has the
added benefit of not necessitating a specific gcc to boot.

We need a patch to create an empty drivers/lego/Kconfig file for the build
to succeed though, as the ev3dev kernel normally expects a drivers tree to
be checked out there.

Signed-off-by: Vincent Stehlé <vincent.stehle@laposte.net>
Suggested-by: Peter Korsgaard <peter@korsgaard.com>