package/edk2-non-osi: new package
Add a package to install the additional edk2 platforms files, not
compatible with the normal licensing requirements and held in the
edk2-non-osi repository.

Only the Marvell Armada files are copied at this point, to support building
edk2 for the MACCHIATObin platform.

The referenced commit corresponds to version edk2-stable202208 of edk2,
based on the timestamps.

This package is heavily inspired from package/edk2-platforms.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>